src/backend_api/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
mangum>=0.17.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
qdrant-client>=1.7.0
openai>=1.10.0
fastembed>=0.2.0
httpx>=0.26.0
#numpy<2.0.0
#ml_dtypes<0.5.0